Game files
The concept is a card game similar to Tinder, where players are presented with cards. If there are four or more negative cards, the display shows red, and if there are four or more positive cards, it shows green. This game idea is designed to challenge players to maintain a balanced approach during the quiz.
- Pritam H